refactor: centralize websocket benchmark env parsing
