+++ packages/runtime/package.json
@@ -13,7 +13,7 @@
     }
   },
   "scripts": {
-    "bench:websocket-firehose": "vp run -t @effect-view-server/config#build && vp run -t @effect-view-server/effect-utils#build && vp run -t @effect-view-server/runtime-core#build && vp run -t @effect-view-server/server#build && compression=$(printf '%s' \"${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_COMPRESSION:-true}\" | sed 's/^[[:space:]]*//;s/[[:space:]]*$//') && if [ -z \"$compression\" ]; then compression=true; fi && case \"$compression\" in true) suffix=-compressed ;; false) suffix= ;; *) printf '%s\\n' 'VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_COMPRESSION must be true or false.' >&2; exit 1 ;; esac && output=${VIEW_SERVER_RUNTIME_BENCH_OUTPUT_JSON:-.artifacts/websocket-firehose-${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_CASE:-same-window}-${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_ROWS:-1000}rows-${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_SUBSCRIBERS:-10}subs${suffix}.json} && mkdir -p .artifacts && VIEW_SERVER_RUNTIME_BENCH_OUTPUT_JSON=$output vp test bench src/websocket-firehose.bench.ts --run --testTimeout 0 --outputJson $output",
+    "bench:websocket-firehose": "vp run -t @effect-view-server/config#build && vp run -t @effect-view-server/effect-utils#build && vp run -t @effect-view-server/runtime-core#build && vp run -t @effect-view-server/server#build && suffix=$(node --input-type=module --eval \"import { booleanFromBenchmarkEnvironment } from '../../scripts/benchmark-environment.mjs'; process.stdout.write(booleanFromBenchmarkEnvironment(process.env['VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_COMPRESSION'], 'VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_COMPRESSION', true) ? '-compressed' : '')\") && output=${VIEW_SERVER_RUNTIME_BENCH_OUTPUT_JSON:-.artifacts/websocket-firehose-${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_CASE:-same-window}-${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_ROWS:-1000}rows-${VIEW_SERVER_RUNTIME_BENCH_WEBSOCKET_SUBSCRIBERS:-10}subs${suffix}.json} && mkdir -p .artifacts && VIEW_SERVER_RUNTIME_BENCH_OUTPUT_JSON=$output vp test bench src/websocket-firehose.bench.ts --run --testTimeout 0 --outputJson $output",
     "build": "vp run -t @effect-view-server/config#build && vp run -t @effect-view-server/effect-utils#build && vp run -t @effect-view-server/runtime-core#build && vp run -t @effect-view-server/server#build && vp pack",
     "test": "node ../../scripts/test-runtime.mjs"
   },
